Make swapchain use RenderTargets
Swapchain creates its own VkImage and VkImageView, which does not play
nicely with the rest of the Framebuffer/RenderTarget set up. This update
makes the swapchain create RenderTargets then sets the properties of it
as a friend. Now they can be added to a framebuffer as if they are any
other kind of RenderTarget.